+++ crates/wdl-grammar/src/experimental/lexer.rs
@@ -226,9 +226,10 @@ pub type LexerResult<T> = Result<T, Error>;
 ///
 /// A lexer produces a stream of tokens from a WDL source string.
 #[allow(missing_debug_implementations)]
+#[derive(Clone)]
 pub struct Lexer<'a, T>(logos::Lexer<'a, T>)
 where
-    T: Logos<'a>;
+    T: Logos<'a, Extras = ()>;
 
 impl<'a, T> Lexer<'a, T>
 where
@@ -269,8 +270,7 @@ where
     /// as the current lexer.
     pub fn morph<T2>(self) -> Lexer<'a, T2>
     where
-        T2: Logos<'a, Source = str, Error = Error>,
-        T::Extras: Into<T2::Extras>,
+        T2: Logos<'a, Source = str, Error = Error, Extras = ()> + Copy,
     {
         Lexer(self.0.morph())
     }
@@ -278,7 +278,7 @@ where
 
 impl<'a, T> Iterator for Lexer<'a, T>
 where
-    T: Logos<'a, Error = Error>,
+    T: Logos<'a, Error = Error, Extras = ()> + Copy,
 {
     type Item = (LexerResult<T>, SourceSpan);