+++ crates/wdl-grammar/CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Fixed
+
+* Fixed parsing of placeholder options in the experimental parser such
+  that it can disambiguate between the `sep` option and a `sep` function
+  call ([#44](https://github.com/stjude-rust-labs/wdl/pull/44)).
+
 ### Added
 
 * Adds support for bound declarations and expressions in the experimental